Move non-webview types to their own contract file
