+++ extension/src/Experiments/index.ts
@@ -6,7 +6,7 @@ import { resolve } from 'path'
 import { ExecutionOptions } from '../cli/execution'
 import { CliReader } from '../cli/reader'
 import { Config } from '../Config'
-import { ExperimentsRepoJSONOutput } from '../Experiments/Webview/contract'
+import { ExperimentsRepoJSONOutput } from '../Experiments/contract'
 import { ExperimentsWebview } from './Webview'
 import { createHash } from 'crypto'
 import { ResourceLocator } from '../ResourceLocator'
